fix: only change necessary properties when setting up svgo config
